% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/int-count_levels.R
\name{count_levels}
\alias{count_levels}
\title{Count Levels
#}
\usage{
count_levels(dat, enum_list, hasNA = c("no", "count.obs", "count.miss"))
}
\arguments{
\item{dat}{A \code{data.frame}. All variables must be factors
#}

\item{enum_list}{A \code{data.frame} consisting of all possible patterns for matching
#with \code{dat}.
#}

\item{hasNA}{A string. Denotes if \code{dat} has complete data or not. If missing
#then should only observed patterns be counted or should missing patterns be counted?}
}
\description{
Given a dataset and a data.frame of comparison patterns,
#count the number of occurances of each pattern.
#
}
\details{
#
}

