% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dep_prior_multi.R
\name{data_dep_prior_multi}
\alias{data_dep_prior_multi}
\title{Data Dependent Prior for Multinomial Distribution}
\usage{
data_dep_prior_multi(dat)
}
\arguments{
\item{dat}{A \code{data.frame}. All variables must be factors}
}
\value{
A \code{data.frame} containing identifiers for all possible \eqn{P(Y=y)} and 
the associated prior-counts, \eqn{\alpha}
}
\description{
Creates a data depedent prior for p-dimensional multinomial distributions
using a conjugate prior (eg \eqn{Dirichlet(\alpha)}) based on 20% of the data.
}
\references{
Darnieder, William Francis. Bayesian methods for data-dependent priors. 
Dissertation. The Ohio State University, 2011.
}
\seealso{
\code{\link{expand.grid}}
}

