% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\docType{class}
\name{imputeMulti-class}
\alias{imputeMulti-class}
\title{Class "imputeMulti" of imputed multinomial data}
\description{
A multivariate multinomial model imputed by EM or Data Augmentation is
represented as a \code{\linkS4class{mod_imputeMulti}} object. A complete
dataset and model is represented as an \code{\linkS4class{imputeMulti}} object.
}
\section{Objects from the class}{
 Objects are created by calls to
\code{\link{multinomial_impute}}, \code{\link{multinomial_em}}, or
\code{\link{multinomial_data_aug}}.
}
\seealso{
\code{\link{multinomial_impute}}, \code{\link{multinomial_em}},
\code{\link{multinomial_data_aug}}
}

