% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_imputeMulti.R
\docType{methods}
\name{show-imputeMulti}
\alias{get_imputations}
\alias{get_imputations,imputeMulti-method}
\alias{n_miss}
\alias{show,imputeMulti-method}
\alias{show-imputeMulti}
\title{Print imputeMulti class objects}
\usage{
\S4method{show}{imputeMulti}(object)

get_imputations(object)

\S4method{get_imputations}{imputeMulti}(object)

n_miss(object)
}
\arguments{
\item{object}{an object of class "imputeMulti"}
}
\description{
print method for class "imputeMulti"

Extract observation level imputed values from imputMulti objects

Extract the number of missing observations prior to imputation from a 'imputeMulti' 
object.
}

