% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\name{is.imputeMulti}
\alias{is.imputeMulti}
\title{Check imputeMulti Class}
\usage{
is.imputeMulti(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
Returns \code{TRUE} if its argument has class "imputeMulti" among its classes and
\code{FALSE} otherwise.
}
\description{
Function that checks if the target object is a \code{imputeMulti} object.
}

