% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\name{is.mod_imputeMulti}
\alias{is.mod_imputeMulti}
\title{Check mod_imputeMulti Class}
\usage{
is.mod_imputeMulti(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
Returns \code{TRUE} if its argument has class "mod_imputeMulti" among its classes and
\code{FALSE} otherwise.
}
\description{
Function that checks if the target object is a \code{mod_imputeMulti} object.
}

