% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_imputed.R
\name{merge_imputed}
\alias{merge_imputed}
\title{Merge imputed data and original dataset}
\usage{
merge_imputed(impute_obj, y, ...)
}
\arguments{
\item{impute_obj}{An object of class "imputeMulti".}

\item{y}{The dataset from which the missing data was imputed.}

\item{...}{Arguments to be passed to other methods}
}
\description{
Merge the imputed dataset from an  \code{imputeMulti} object with the original dataset. 
Merging is done by rownames, since imputeMulti maintains row-order during imputation.
}

