% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\docType{class}
\name{mod_imputeMulti-class}
\alias{mod_imputeMulti-class}
\title{Class "mod_imputeMulti"}
\description{
A multivariate multinomial model imputed by EM or Data Augmentation is
represented as a \code{\linkS4class{mod_imputeMulti}} object. A complete
dataset and model is represented as an \code{\linkS4class{imputeMulti}} object.
Slots for \code{mod_imputeMulti} objects include: (1) the modeling method;
(2) the call to the estimation function; (3) the number of iterations in estimation;
(4) the final log-likelihood; (5) the conjugate prior if any; (6) the MLE estimate of
the sufficient statistics and parameters.
}
\section{Objects from the class}{
 Objects are created by calls to
\code{\link{multinomial_impute}}, \code{\link{multinomial_em}}, or
\code{\link{multinomial_data_aug}}.
}
\seealso{
\code{\link{multinomial_impute}}, \code{\link{multinomial_em}},
\code{\link{multinomial_data_aug}}
}

