% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_imputeMulti.R
\docType{methods}
\name{show-mod_imputeMulti}
\alias{get_iterations}
\alias{get_iterations,mod_imputeMulti-method}
\alias{get_logLik}
\alias{get_logLik,mod_imputeMulti-method}
\alias{get_method}
\alias{get_method,mod_imputeMulti-method}
\alias{get_parameters}
\alias{get_parameters,mod_imputeMulti-method}
\alias{get_prior}
\alias{get_prior,mod_imputeMulti-method}
\alias{n_miss,imputeMulti-method}
\alias{show,mod_imputeMulti-method}
\alias{show-mod_imputeMulti}
\title{Print mod_imputeMulti class objects}
\usage{
\S4method{show}{mod_imputeMulti}(object)

get_parameters(object)

\S4method{get_parameters}{mod_imputeMulti}(object)

get_prior(object)

\S4method{get_prior}{mod_imputeMulti}(object)

get_iterations(object)

\S4method{get_iterations}{mod_imputeMulti}(object)

get_logLik(object)

\S4method{get_logLik}{mod_imputeMulti}(object)

get_method(object)

\S4method{get_method}{mod_imputeMulti}(object)

\S4method{n_miss}{imputeMulti}(object)
}
\arguments{
\item{object}{an object of class "mod_imputeMulti"}
}
\description{
print method for class "mod_imputeMulti"

Extract parameter estimates from a 'mod_imputeMulti' object

Extract the character string specifying the prior from a 'mod_imputeMulti' object

Extract the number of iterations used in fitting a 'mod_imputeMulti' object

Extract the final log likelihood from fitting a 'mod_imputeMulti' object

Extract the method used in fitting a 'mod_imputeMulti' object. Returns either 
EM (expectation-maximization) or DA (data-augmentation).
}

