% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multinomial_em.R
\name{multinomial_em}
\alias{multinomial_em}
\title{EM algorithm for multinomial data}
\usage{
multinomial_em(x_y, z_Os_y, enum_comp, n_obs, conj_prior = c("none",
  "data.dep", "flat.prior", "non.informative", "select"), alpha = NULL,
  tol = 1e-08, max_iter = 10000, verbose = FALSE)
}
\arguments{
\item{x_y}{A \code{data.frame} of observed counts for complete observations.}

\item{z_Os_y}{A \code{data.frame} of observed marginal-counts for incomplete observations.}

\item{enum_comp}{A \code{data.frame} specifying a vector of all possible observed patterns.}

\item{conj_prior}{A string specifying the conjugate prior. One of
\code{c("none", "data.dep", "flat.prior", "non.informative", "select")}.}

\item{alpha}{The vector of counts \eqn{\alpha} for a \eqn{Dir(\alpha)} prior. Must be specified if
\code{conj_prior} is either \code{c("data.dep", "flat.prior")}. If \code{flat.prior}, specify
as a scalar. If \code{data.dep}, specify as a vector with key matching \code{enum_comp}.}

\item{tol}{A scalar specifying the convergence criteria. Defaults to \code{1e-8}}

\item{max_iter}{An integer specifying the maximum number of allowable iterations. Defaults
to \code{10000}.}

\item{verbose}{Logical. If \code{TRUE}, provide verbose output on each iteration.}
}
\value{
A \code{list} containing the following: (1) The function call; (2) the
number of iterations; (3) The observed data log likelihood; (4) the MLE conjugate prior
 (if requested); (5) the ML estimates from EM.
}
\description{
Implement the EM algorithm for multvariate multinomial data given
observed counts of complete and missing data (Y_obs and Y_mis). Allows for specification
of a Dirichlet conjugate prior.
}

