% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\docType{methods}
\name{show,imputeMulti-method}
\alias{show,imputeMulti-method}
\title{Print imputeMulti class objects}
\usage{
\S4method{show}{imputeMulti}(object)
}
\arguments{
\item{object}{an object of class "imputeMulti"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print method for class "imputeMulti"
}

