% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\docType{methods}
\name{show,mod_imputeMulti-method}
\alias{show,mod_imputeMulti-method}
\title{Print mod_imputeMulti class objects}
\usage{
\S4method{show}{mod_imputeMulti}(object)
}
\arguments{
\item{object}{an object of class "mod_imputeMulti"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print method for class "mod_imputeMulti"
}

