% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_imputeMulti.R
\docType{methods}
\name{summary,imputeMulti-method}
\alias{summary,imputeMulti-method}
\title{Summarizing imputMulti objects}
\usage{
\S4method{summary}{imputeMulti}(object, ...)
}
\arguments{
\item{object}{an object of class "imputeMulti"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary method for class "imputeMulti"
}

