% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_imputeMulti.R
\docType{methods}
\name{summary,mod_imputeMulti-method}
\alias{summary,mod_imputeMulti-method}
\title{Summarizing mod_imputMulti objects}
\usage{
\S4method{summary}{mod_imputeMulti}(object, ...)
}
\arguments{
\item{object}{an object of class "mod_imputeMulti"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary method for class "mod_imputeMulti"
}

