% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\name{summary.imputeMulti}
\alias{summary.imputeMulti}
\title{Summarizing imputMulti objects}
\usage{
summary.imputeMulti(object, ...)
}
\arguments{
\item{object}{an object of class "imputeMulti"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary method for class "imputeMulti"
}

