% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_imputeMulti.R
\name{summary.mod_imputeMulti}
\alias{summary.mod_imputeMulti}
\title{Summarizing mod_imputMulti objects}
\usage{
summary.mod_imputeMulti(object, ...)
}
\arguments{
\item{object}{an object of class "mod_imputeMulti"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
summary method for class "mod_imputeMulti"
}

