% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{supDistC}
\alias{supDistC}
\title{Calculate the sup of L1 distance between x and y}
\usage{
supDistC(x, y)
}
\arguments{
\item{x}{A numeric \code{vector}}

\item{y}{A numeric \code{vector}}
}
\value{
a numeric scalar.
}
\description{
sup of L1 distance between x and y
}
