% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{xy_compare}
\alias{xy_compare}
\title{Compare two integer matrices, allowing missing values}
\usage{
xy_compare(mat_x, mat_y)
}
\arguments{
\item{mat_x}{A two dimensional array which may contain missing values}

\item{mat_y}{A two dimensional array without missing values}
}
\value{
A \code{list} of matches.
}
\description{
Compare two two-dimensional arrays (\code{mat_x}, \code{mat_y}), where \code{mat_x}
permits missing values. Return a \code{list} of length \code{nrow(mat_x)} such that each list
element contains a vector of row indices from \code{mat_y} with row-equivalence of the non
missing values.
}
