% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multinomial_impute.R
\name{multinomial_impute}
\alias{multinomial_impute}
\title{Impute Values for missing multinomial values}
\usage{
multinomial_impute(dat, method = c("EM", "DA"), conj_prior = c("none",
  "data.dep", "flat.prior", "non.informative"), alpha = NULL,
  verbose = FALSE, tol = 5e-07)
}
\arguments{
\item{dat}{A \code{data.frame}. All variables must be factors.}

\item{method}{\code{c("EM", "DA")} A string specifying EM or Data Augmentation (DA)}

\item{conj_prior}{A string specifying the conjugate prior. One of
\code{c("none", "data.dep", "flat.prior", "non.informative")}.}

\item{alpha}{The vector of counts \eqn{\alpha} for a \eqn{Dir(\alpha)} prior. Must be specified if
\code{conj_prior} is either \code{c("data.dep", "flat.prior")}. If \code{flat.prior}, specify
as a scalar. If \code{data.dep}, specify as a vector with key matching \code{enum_comp}.}

\item{verbose}{Logical. If \code{TRUE}, provide verbose output on each iteration.
#'}

\item{tol}{A scalar specifying the convergence criteria. Defaults to \code{5e-7}}

\item{alpha}{The vector of counts \eqn{\alpha} for a \eqn{Dir(\alpha)} prior. Must be specified if
\code{conj_prior} is either \code{c("data.dep", "flat.prior")}. If \code{flat.prior}, specify
as a scalar. If \code{data.dep}, specify as a vector with key matching \code{enum_comp}.}
}
\description{
Impute values for multivariate multinomial data using either EM or Data
Augmentation.
}
\references{
Schafer, Joseph L. Analysis of incomplete multivariate data. Chapter 7.
CRC press, 1997.
}
\seealso{
\code{\link{expand.grid}}, \code{\link{data_dep_prior_multi}}, \code{\link{multinomial_em}}
}

